chore(sdk): clean configs
