+++ apps/sdk/src/lib.rs
@@ -6,13 +6,11 @@ use tokio::sync::RwLock;
 
 pub mod client;
 pub mod command;
-pub mod config;
 pub mod message;
 pub mod stream;
 
 pub use client::NodeClient;
 pub use command::{BotCommand, COMMAND_HINT, parse_command};
-pub use config::{TelegramConfig, WechatConfig};
 pub use message::{Attachment, AttachmentKind, GatewayMessage, attachment_summary};
 pub use stream::StreamAccumulator;
