+++ apps/telegram/Cargo.toml
@@ -20,6 +20,8 @@ command.workspace = true
 
 anyhow.workspace = true
 clap.workspace = true
+serde.workspace = true
+toml.workspace = true
 crossterm.workspace = true
 futures-util.workspace = true
 serde_json.workspace = true
