+++ apps/telegram/src/serve.rs
@@ -4,7 +4,7 @@ use crate::{
     COMMAND_HINT, GatewayMessage, KnownBots, NodeClient, StreamAccumulator, StreamResult,
     attachment_summary, parse_command,
 };
-use sdk::config::TelegramConfig;
+use crate::config::TelegramConfig;
 use std::{collections::HashMap, sync::Arc};
 use teloxide::prelude::*;
 use teloxide::types::{ChatAction, InlineKeyboardButton, InlineKeyboardMarkup};
